+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Fixes
 
 - Cache network capabilities and status to reduce IPC calls ([#4560](https://github.com/getsentry/sentry-java/pull/4560))
+- Deduplicate battery breadcrumbs ([#4561](https://github.com/getsentry/sentry-java/pull/4561))
+
 
 ## 8.18.0